requirements.txt
transformers
torch
sentence-transformers
pandas
numpy
scikit-learn
tqdm
faiss-cpu

#preprocessing
pandas
numpy
faiss-cpu
sentence-transformers
torch